Replace header symlink with environment variable.
Symlinks in repos are problematic for a number of reasons (not supported on
windows, can't be installed with qlot), and are a little cumbersome to
change. Rather than a symlink stored in the repo, use an environment
variable to set the location of headers. This defaults to /usr/include for
convenience on unices, where there is a common header location.

Since the business about needing to include the parent directory of the
sodium/ header directory is easy to get mixed up, this includes a test for
a header file to print an informative message if it looks like we can't
locate the headers with the given config.