Use common sodium.h header to create bindings.
Newer versions of libsodium ship with a single common header that includes
all the other headers in the library. SWIG can't process such a header
itself (it doesn't recurse into includes), but with a little preprocessing
we can construct an interface file on-the-fly that will process all of the
headers included by the common header. This should avoid future issues with
headers that are removed, and also eliminate the burden of having to update
the bindings file every time libsodium adds new algorithms.

Note that the existing binding file is left in place and used if the common
header isn't found, to support older versions. The common header also uses
an oddball '# include' directive at present, so we have to match spaces in
the include directive during preprocessing.