Fix header ordering issue.
crypto_secretstream_xchacha20poly1305.h uses constants defined by
crypto_stream_chacha20.h, but that header is processed after the former
in the SWIG interface, so SWIG generates a DEFSTRUCT form that uses a
constant symbol into its value. At a minimum we would need to fix this up
to lispify the name, but the constant winds up being defined later than its
use, and this can't be loaded at all.

(This is because DEFSTRUCT uses the count parameter at macroexpansion time,
so in order to use a constant symbol you must use #., and unlike an
ordinary reference, that can't be deferred to the end of the compilation
unit where the constant will actually be defined.)

Instead, make sure we process crypto_stream_chacha20.h first, the hard way,
by inlining it into the top of the interface like we do for export.h.