Fix length-suffixed integers.
If the C code includes integers with a length suffix (like '32u'), SWIG
will include the suffix in the integer code it generates, which makes it an
undefined symbol in lisp, rather than an integer as intended. The suffix
is now stripped after generation, being careful to include do this with
integer-looking things followed by whitespace or a close-paren (this is
safer than it might seem: in C an identifier token cannot begin with an
integer, so these are only going to be numeric sorts of things).