Add a clarifying comment to the (now) fallback interface file.
We don't want debugging to be harder than it has to be, and it would be
very easy for somebody to see an issue and pop the bindings file open to
see what wrong (or to make a change there thinking that was the interface
used to generate bindings in most cases). To avoid confusion, add a big,
visible comment to the bindings file that is now only a fallback for old
libsodium versions.